+++ src/control/ConfigurationControl.java
@@ -2,8 +2,8 @@ package control;
 
 import org.bson.types.ObjectId;
 
-import dao.EmployeeDAO;
-import dao.EmployeeDAOImpl;
+import dao.impl.EmployeeDAOImpl;
+import dao.interfaces.EmployeeDAO;
 import entity.Employee;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
