+++ src/dao/impl/ParametersDAOImpl.java
@@ -1,29 +1,23 @@
-package dao;
-
-import java.time.LocalDate;
+package dao.impl;
 
 import com.mongodb.client.MongoCollection;
-
+import dao.interfaces.ParametersDAO;
+import java.time.LocalDate;
 import org.bson.Document;
 import org.bson.types.ObjectId;
-
 import utils.MongoConnect;
 
 public class ParametersDAOImpl implements ParametersDAO {
 
   MongoCollection<Document> parameters;
+  MongoConnect mc = new MongoConnect();
 
-  public ParametersDAOImpl() {
-		connection();
-	}
-
-  void connection() {
-    MongoConnect mc = new MongoConnect();
-    mc.connection();
+  void getCollection() {
     parameters = mc.database.getCollection("parameters");
   }
 
   public String dailyPhrase() {
+    getCollection();
     Document phrase = parameters
       .find(new Document("day", LocalDate.now().getDayOfMonth()))
       .first();
@@ -31,15 +25,16 @@ public class ParametersDAOImpl implements ParametersDAO {
   }
 
   public boolean insertDailyPhrase(String phrase, int day) {
-		try {
-			parameters.insertOne(
-				new Document("_id", new ObjectId())
-					.append("day", day)
-					.append("phrase", phrase)
-			);
-		} catch (Exception e) {
-			return false;
-		}
-		return true;
+    getCollection();
+    try {
+      parameters.insertOne(
+        new Document("_id", new ObjectId())
+          .append("day", day)
+          .append("phrase", phrase)
+      );
+    } catch (Exception e) {
+      return false;
+    }
+    return true;
   }
 }
