+++ src/utils/MongoConnect.java
@@ -1,7 +1,5 @@
 package utils;
 
-import java.util.logging.Logger;
-
 import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientSettings;
 import com.mongodb.client.MongoClient;
@@ -11,23 +9,23 @@ import com.mongodb.client.MongoDatabase;
 public class MongoConnect {
 
   AuthMDB auth = new AuthMDB();
-	DBLogger lgr = new DBLogger();
-	public MongoDatabase database;
-	
-	public void connection() {
-		lgr.setupLogger();
-		ConnectionString connectionString = new ConnectionString(
-			"mongodb+srv://" +
-			auth.getAuth() +
-			"@clusterpetsanatomy.pxsfd.mongodb.net/?retryWrites=true&w=majority"
-		);
-	
-		MongoClientSettings settings = MongoClientSettings
-			.builder()
-			.applyConnectionString(connectionString)
-			.build();
-	
-		MongoClient mongoClient = MongoClients.create(settings);
-		database = mongoClient.getDatabase("PetsAnatomy");
-	}
+  DBLogger lgr = new DBLogger();
+  public MongoDatabase database;
+
+  public void connection() {
+    lgr.setupLogger();
+    ConnectionString connectionString = new ConnectionString(
+      "mongodb+srv://" +
+      auth.getAuth() +
+      "@clusterpetsanatomy.pxsfd.mongodb.net/?retryWrites=true&w=majority"
+    );
+
+    MongoClientSettings settings = MongoClientSettings
+      .builder()
+      .applyConnectionString(connectionString)
+      .build();
+
+    MongoClient mongoClient = MongoClients.create(settings);
+    database = mongoClient.getDatabase("PetsAnatomy");
+  }
 }
